m-tex-live-utility: add livecheck
This adds a `livecheck` block for `m-tex-live-utility`, as the cask
`url` is a GitHub release asset but livecheck is using the `Git`
strategy by default. Like `m-obs`, this cask uses both stable and
unstable versions, so this adds a `GithubReleases` `strategy` block
that only avoids draft releases (instead of both draft and
pre-release releases) and simply uses the full tag name for each
release.